Create Menus for all of the current game states
This refactor generalizes the menu system to all game states and
initializes all of them. This simplifies the update loop, since we can
just call the current active menu for screen writing. Additionally,
active menus are now notified via the UpdateState() function when game
mode changes, because isActive==false. This is used in the PlayScreen
to reset the drawing mode fill color when the mode switches.