Add TextLayout utility
This commit adds a new utility/ folder to sources/ to keep them
away from the menu code. It also introduces a new TextLayout module
that defines the various rows/columns text should be aligned to across
the game. It's the bare minimum a "textlayout" helper should do and
there's room to grow as text layout needs increase in the project.